\name{hoopoe1}
\alias{hoopoe1}
\docType{data}
\title{
Light intensity measurements over time recorded on a migratory bird
}
\description{
Sunlight intensity measurements over time recorded during the first part of the annual migration of a European Hoopoe (\cite{Upupa epops}). All dates/times are measured in Universal Time Zone (UTC).
}
\usage{data(hoopoe1)}
\format{
A table with 24474 rows and 2 columns, rows corresponding to light measurements recorded in ten-minute intervals (datetime).
}
\source{
Baechler, E., Hahn, S., Schaub, M., Arlettaz, R., Jenni, L., Fox, J.W., Afanasyev, V. & Liechti, F. (2010) Year-Round Tracking of Small Trans-Saharan Migrants Using Light-Level Geolocators. \emph{Plos One}, \bold{5}.
}