% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLight.R
\name{i.argCheck}
\alias{i.argCheck}
\title{Checkup of arguments in GeoLight tables}
\usage{
i.argCheck(y)
}
\arguments{
\item{y}{GeoLigth data table.}
}
\description{
Checkup of arguments in GeoLight tables
}
\author{
Simeon Lisovski
}
