% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLight.R
\name{ligTrans}
\alias{ligTrans}
\title{Transformation of *.lig files}
\usage{
ligTrans(file)
}
\arguments{
\item{file}{the full patch and filename with suffix of the *.lux file.}
}
\value{
A \code{data.frame} suitable for further use in
\bold{\code{GeoLight}}.
}
\description{
Transform *.lig files derived from geolocator
deviced for further analyses in \bold{\code{GeoLight}}.
}
\seealso{
\code{\link{gleTrans}} for transforming *.glf files produced by the
software GeoLocator (\emph{Swiss Ornithological Institute})
}
\author{
Tamara Emmenegger
}
