% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeoLight.R
\name{mergeSites}
\alias{mergeSites}
\title{Function to merge sites}
\usage{
mergeSites(datetime, crds, site, threshold = 250, cutoff = 0.6,
  plot = TRUE...)
}
\arguments{
\item{datetime}{vector of the dates that correspond to the positions (in .POSIX class).}

\item{crds}{a \code{SpatialPoints} or \code{matrix} object, containing x
and y coordinates (in that order).}

\item{site}{a \code{numerical vector} assigning each row to a particular
period. Stationary periods in numerical order and values >0,
migration/movement periods 0. This \code{vector} will be used as the initial state.}

\item{threshold}{a \code{numerical} value defining the threshold of the distance under
which consequtive sites should be merged (in km).}

\item{cutoff}{a \code{numerical} value indicating the percentile of the distances matrix that
will be used to merge sites based on the defined \code{treshold}.}

\item{plot}{\code{logical}, if TRUE a plot comparing the inital and the final site selection.}
}
\value{
A \code{vector} with the merged site numbers
}
\description{
The \code{\link{changeLight}} functions provides a vector grouping the twilight times
into stationary (>0) and movement (0) periods. This function was written to enable the user
to merge sites based on the distance between consequtive sites. NOTE: The function requires
position estimate and desicison on whether sites should be merged will be made based on
the defined \code{distance}, the \code{cutoff} values and the provided positions. The analysis
is this dependent on the accuracy of the position estiamtes and should be applied to positons that
were estimated using a sensible sun elevation angle.
}
\author{
Simeon Lisovski
}
\seealso{
\code{\link{changeLight}}
}

