% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLight.R
\name{siteEstimate}
\alias{siteEstimate}
\title{siteEstimate}
\usage{
siteEstimate(
  tFirst,
  tSecond,
  type,
  twl,
  degElevation,
  method = "gamma",
  parms = c(3.3, 0.8),
  xlim = c(-180, 180),
  ylim = c(-90, 90),
  res = c(0.5, 0.5)
)
}
\arguments{
\item{tFirst}{vector of sunrise/sunset times (e.g. 2008-12-01 08:30).}

\item{tSecond}{vector of of sunrise/sunset times (e.g. 2008-12-01 17:30).}

\item{type}{vector of either 1 or 2, defining \code{tFirst} as sunrise or sunset respectively.}

\item{twl}{data.frame containing twilights and at least \code{tFirst}, \code{tSecond} and \code{type}}

\item{degElevation}{the sun elevation angle (in degrees) that defines twilight (e.g. -6 for "civil
twilight"). Either a single value, a \code{vector}.}

\item{method}{\code{character} string; only \code{gamma} and \code{log-normal} are implemented.}

\item{parms}{a \code{vector} describing the two parameters of the error density distribution (defined by \code{method}).}

\item{xlim}{the longitudinal boundaries for which the likelihood will be calculated.}

\item{ylim}{the latitudinal boundaries for which the likelihood will be calculated.}

\item{res}{the spatial resolution in degrees.}
}
\value{
A \code{list} with ...
}
\description{
...
}
\author{
Simeon Lisovski
}
