% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeoLight.R
\name{staroddiTrans}
\alias{staroddiTrans}
\title{Transformation of staroddi files}
\usage{
staroddiTrans(file)
}
\arguments{
\item{file}{the full patch and filename of the staroddi file.}
}
\value{
A \code{data.frame} suitable for further use in
\bold{\code{GeoLight}}.
}
\description{
Transform staroddi files derived from geolocator
deviced for further analyses in \bold{\code{GeoLight}}.
}
\author{
Tamara Emmenegger
}

