% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeoLight.R
\name{trnTrans}
\alias{trnTrans}
\title{Transformation of *.trn files}
\usage{
trnTrans(file)
}
\arguments{
\item{file}{the full patch and filename of the *.trn file.}
}
\value{
A \code{data.frame} suitable for further use in
\bold{\code{GeoLight}}.
}
\description{
Transform *.trn files derived from geolocator
deviced for further analyses in \bold{\code{GeoLight}}.
}
\author{
Tamara Emmenegger
}

