% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLight.R
\name{getElevation}
\alias{getElevation}
\title{Calculate the appropriate sun elevation angle for known location}
\usage{
getElevation(
  tFirst,
  tSecond,
  type,
  twl,
  known.coord,
  method = "log-norm",
  plot = TRUE
)
}
\arguments{
\item{tFirst}{vector of sunrise/sunset times (e.g. 2008-12-01 08:30).}

\item{tSecond}{vector of of sunrise/sunset times (e.g. 2008-12-01 17:30).}

\item{type}{vector of either 1 or 2, defining \code{tFirst} as sunrise or sunset respectively.}

\item{twl}{data.frame containing twilights and at least \code{tFirst}, \code{tSecond} and \code{type} (alternatively give each parameter separately).}

\item{known.coord}{a \code{SpatialPoint} or \code{matrix} object, containing
known x and y coordinates (in that order) for the selected measurement
period.}

\item{method}{the function can either estimate the sun elevation angle and the twilight error parameters using a log-normal ("log-norm")
or a gamma ("gamma") error distribution. It is recommended to try both and evaluate the fit using the plot.}

\item{plot}{\code{logical}, if TRUE a plot will be produced.}
}
\value{
A \code{data.frame} containing the sun elevation anhle (0), the Sun elevation angle (mean), shape and scale
}
\description{
Function to calculate the sun elevation angle for light measurements at a
known location and the choosen light threshold.
}
\details{
NEW: The function provides two different sun elevation angle. The first (a1) refers to the median twiligth error and
is needed for threshold based locaiton estimation (e.g. \code{\link{coord}}). The second (a0) refers to the zero elevation
angle of the twilight error distribution and is required for e.g. \code{\link{mergeSites2}}, \code{\link{siteEstimate}}. The function
also provides the parameters (log.mean and log.sd) of the fitted log-normal distribution (see red line in plot),
}
\examples{
data(calib2)
  calib2$tFirst  <- as.POSIXct(calib2$tFirst, tz = "GMT")
  calib2$tSecond <- as.POSIXct(calib2$tSecond, tz = "GMT")
getElevation(calib2, known.coord = c(7.1,46.3))
}
\references{
Lisovski, S., Hewson, C.M, Klaassen, R.H.G., Korner-Nievergelt,
F., Kristensen, M.W & Hahn, S. (2012) Geolocation by light: Accuracy and
precision affected by environmental factors. \emph{Methods in Ecology and
Evolution}, DOI: 10.1111/j.2041-210X.2012.00185.x.
}
\author{
Simeon Lisovski
}
